Implement & document validity checks
Implement and document a plugin that removes products that contain
insufficient valid data

Remove any products where the fraction valid_data/expected_valid_data is
less than a configured threshold in %.  Expected valid data is calculated
by the scene coverage for each resampled scene.  This plugin was designed
for use with AVHRR, which may alternate between channels 3A and 3B.
Since this is different between resampled scenes, this plugin must be
applied after scene resampling.

This will trigger a calculation for the data to be checked.

To be configured with the ``rel_valid`` key indicating validity in %.
For example:

    product_list:
      areas:
	fribbulus_xax:
	  red:
	    min_valid: 40

    workers:
      - fun: !!python/name:trollflow2.plugins.create_scene
      - fun: !!python/name:trollflow2.plugins.load_composites
      - fun: !!python/name:trollflow2.plugins.resample
      - fun: !!python/name:trollflow2.plugins.check_valid
      - fun: !!python/name:trollflow2.plugins.save_datasets