Fix scene content checking
Fix scene content checking checking that a product is in a scene, not in
scene.keys(), as the latter has different behaviour.  This wasn't caught
by the unit test which uses a scene as a dict, because scenes actually
producte DataIDs when keys are set.  Therefore, use a real Scene object
here (this means the unit tests will need satpy).