Remove coverage per product.  It's too complicated.
Delete remnants of my attempt to estimate coverage per product.  It's
too complicated.  Unlike regular covers(), it needs to be done after
loading the datasets, because the dataset attributes contain necessary
information on start and end times.  But the start and end times are not
updated when resampling.  There are also corner cases on what to do with
products that don't have start/end time or that weren't loaded.  And in
the end, it wouldn't fully solve the problem that filtering does either,
because the data source isn't always thruthful about what channel is
switched on.