Move import to be dask-dist friendly
Make trollflow2 more multiprocessing / dask-distributed friendly.
Move the launcher import in satpy_launcher.py within the main
function, so that it only gets executed in the primary process.
The call to `Manager()` in `trollflow2.__init__.py` was causing
attempts to use trollflow2 with dask distributed to fail with a
RuntimeError.